Improve handling of voxel bounds
Adds a struct `VoxelBounds` containing arrays for the lower and
upper bounds based on voxel IDs. This also simplifies the
signature of of the function to obtain the voxel bounds. The function
for obtaining these bounds was renamed to `get_voxel_bounds` to
better reflect the functions purpose and actions.