Handle coordinate transform in classes
This prepares the use of dedicated classes to convert between our
internal coordinate systems (index, annotation and CDP) to the voxel
coordinate system of VDS.

The classes offer automatic conversion between the request coordinate
system and the voxel coordinate system via class overloading. This
separation makes it more clear what is expected in each of the internal
coordinate systems.