Add an reverse proxy for http caching
Warning: This is very experimental

This commit does two things, and thus should in theory be 2 commits.

A new program - the httpcache
-----------------------------

The httpcache is implemented as a reverse proxy. A cache header
("If-Modified-Since") will be added to incomming requests before
forwarding, if the request already exists in cache. The value is set to
match the "Last-Modified" header from the cached response. This act both
as authentication and cache invalidation.

Responses are intercepted and amended based on the status code. A
NotModified (304) means it's safe to return what we have in cache. A 200
response will be cached before returning. And everyhting else is just a
straight pass-through.

Routing OpenVDS traffic through the httpcache
---------------------------------------------

Using OpenVDS' http iodevice we can alter the connection string such
that OpenVDS sends it's blob request towards our proxy. The host and
blob part is encoded in the url part. Query parameters, headers, body
and additional paths that OpenVDS sets up will be perserved and
forwarded to the proxy and all they way to it's destination.

How it ties together
--------------------

 +------+         +--------+         +------------+         +------------+
 | user | --(1)-> | server | --(2)-> | http cache | --(3)-> | blob store |
 |      | <-(6)-- |        | <-(5)-- |            | <-(4)-- |            |
 +------+         +--------+         +------------+         +------------+

 Performance
 -----------

 This is suprisingly slow. On a request that avarage out at ~10 sec
 (140MB fence), it takes ~7.5 sec to fetch it, even when everything is
 in cache.

 It is true we do exactly the same number of requests to blob store
 regarless of we hit cache or not. But I thought that the difference
 between blob store returning empty requests (blob not modified since
 last access) and returning the actual blob would be significant. Looks
 like I'm wrong. It would be interesting to see how this performs on
 different brick sizes and compressions (i.e. how large the
 blobs/downloads are)