Add new radix env 'grpc'
The env runs 5 component, one scheduler and four worker. The scheduler
and workers communicate through gRPC. See the upstream branch grpc for
implementation details

The gRPC workers are currently set up as 4 separate components. It could
also be a single component with 4 replicas, but then we rely on radix'
load balancing to distribute work. That might be fine, but for testing
we've preferred to use the gRPC's own load balancing. Which is the one
we've been testing locally.