Add subslicing support to the API
This commit extends the slicing API to allow requests for server side
slice cropping. A set of axis bounds can be provided in the request in
order to get a subslice. Slices can be cropped in one or both
directions, or none of them. Like the slice itself, bounds can be
specified in either the index or annotation coordinate system.

Bounds are optional and extends the public API without breaking it.